vandalizer-next/backend/pyproject.toml
[project]
name = "vandalizer-next-backend"
version = "0.1.0"
description = "Vandalizer Next - FastAPI backend"
requires-python = ">=3.11,<3.13"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "beanie>=1.27",
    "motor>=3.6",
    "pydantic-settings>=2.7",
    "python-jose[cryptography]>=3.3",
    "werkzeug>=3.1",
    "python-multipart>=0.0.18",
    "celery[redis]>=5.4",
    "pydantic-ai>=0.1",
    "pymongo>=4.0",
    "chromadb>=0.5",
    "langchain-text-splitters>=0.3",
    "langchain-openai>=0.3",
    "httpx>=0.27",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
